requirements.txt
requests
azure-identity
azure-keyvault-secrets
botbuilder-core
openai